+++ src/Treehouse/Treehouse.Database/Models/Job.cs
@@ -11,7 +11,7 @@ namespace TreeHouse.Database.Models
         [Required]
         public string Name { get; set; }
         public string Description { get; set; }
-        public double? Amount { get; set; }
+        public decimal? Amount { get; set; }
         public int UserId { get; set; }
         public User User { get; set; }
         public List<JobComment> Comments { get; set; }
